requirements.txt
groq
pydantic
python-dotenv
fastapi
uvicorn
jupyter
